+++ database/migrations/2022_12_05_071558_tbl_bahan_keluar.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('kd_bahan', 10);
             $table->string('nm_bahan', 50);
             $table->date('tgl_keluar');
-            $table->integer('jumlah');
+            $table->float('jumlah');
             $table->double('total');
             $table->string('ket');
             $table->timestamps();
